+++ changelog.markdown
@@ -97,3 +97,4 @@
 * Replaced calls to **error** with **[]**, in implementation of **Read**, in module **RegExChar.ExtendedRegExChar**.
 * Added file **.travis.yml** to control testing by <https://docs.travis-ci.com>.
 * Added file **.ghci**.
+* Replaced use of module **ToolShed.Defaultable** with **Data.Default**.